backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangGraph and AI
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.0
langchain-core==0.2.0
openai==1.3.7

# Research and Data Processing
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0 